+++ packages/terra-table/src/Table.jsx
@@ -106,6 +106,12 @@ const propTypes = {
   rowHeight: PropTypes.string,
 
   /**
+   * A string that specifies the Minimum height for the rows on the table. rowHeight takes precedence if valid CSS value is passed.
+   * With this property the height of the cell will grow to fit the cell content.
+   */
+  rowMinimumHeight: PropTypes.string,
+
+  /**
    * A number indicating the index of the column that represents the row header. The index is based on 0 and cannot exceed one less than the number of columns on the table.
    */
   rowHeaderIndex: validateRowHeaderIndex,
@@ -176,7 +182,7 @@ const defaultProps = {
   rowHeaderIndex: 0,
   defaultColumnWidth: 200,
   columnHeaderHeight: '2.5rem',
-  rowHeight: '2.5rem',
+  rowMinimumHeight: '2.5rem',
   pinnedColumns: [],
   overflowColumns: [],
   rows: [],
@@ -212,6 +218,7 @@ function Table(props) {
     isStriped,
     rowHeaderIndex,
     intl,
+    rowMinimumHeight,
   } = props;
 
   // Manage column resize
@@ -603,6 +610,7 @@ function Table(props) {
               rowHeaderIndex={rowHeaderIndex}
               onCellSelect={isGridContext || rowSelectionMode ? handleCellSelection : undefined}
               onSectionSelect={onSectionSelect}
+              rowMinimumHeight={rowMinimumHeight}
             />
           ))}
         </ColumnContext.Provider>